requirements.txt
pandas
datasets
llama-cpp-python
ollama
torch
transformers
tqdm
scikit-learn
sctokenizer
kagglehub
faiss-cpu # use faiss-gpu maybe
matplotlib
mlxtend
